src/requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Django==3.2.3
django-debug-toolbar==3.2.1
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-extensions==3.1.3
django-nine==0.2.4
django-pandas==0.6.4
djangorestframework==3.12.4
docopt==0.6.2
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
Flask==2.0.1
flask-restplus==0.13.0
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.20.3
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.9.0
pymdown-extensions==8.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
serializers==0.2.4
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.5
Werkzeug==2.0.1
